+++ apps/formio-server/src/actions/googlesheet/googleSheet.js
@@ -22,10 +22,6 @@ module.exports = (router) => {
    *  This class is used to create the Google Sheet action.
    */
   class GoogleSheetAction extends Action {
-    constructor(data, req, res) {
-      super(data, req, res);
-    }
-
     static info(req, res, next) {
       next(null, hook.alter('actionInfo', {
         name: 'googlesheet',