+++ apps/formio-server/Changelog.md
@@ -6,7 +6,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## 6.10.4-rc.1
 ### Changed
+ - Upgrade dependencies
  - Upgrade portal@7.0.36-rc.1
+ - Upgrade formiojs@4.11.0-rc.2
+ - Upgrade formio@1.89.0
+
+### Fixed
+ - FOR-2682: Fixes an issue where x-jwt-token from SAML was too big
+ - fix: added next function calls
 
 ## 6.10.3
 ### Changed